% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCA_viz.R
\name{anchor}
\alias{anchor}
\title{anchor()
Question: should it be moved to a different package?
threshholds function
this function allows to randomize the threshholds
for each anchor uploaded with the data by generating
a 10% mark-ups around the threshholds}
\usage{
anchor(a, b, c)
}
\arguments{
\item{a}{First threshold}

\item{b}{Second threshold}

\item{c}{Third threshhold}
}
\value{
three numbers which are randomly generated within
10% range of the initial input numbers
}
\description{
anchor()
Question: should it be moved to a different package?
threshholds function
this function allows to randomize the threshholds
for each anchor uploaded with the data by generating
a 10% mark-ups around the threshholds
}
\note{
It is important to remember that random number
draws are obtained from a uniform distribution
}
