% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCA_viz.R
\name{comparison}
\alias{comparison}
\title{Main Functions
comparison()}
\usage{
comparison(x = all_values, y, num = F)
}
\arguments{
\item{x}{list of all the solutions or configurations
which are obtained from multiple QCA solutions}

\item{y}{QCA solutions in their raw form as produced
by \pkg{QCA} package}

\item{num}{is a default argument which checks
whether the inputs are numeric or not. The default
is set for false. This ensures that even if the input
is non-numeric, the function will read in the
data in a numeric format.}
}
\value{
The function counts the individual solutions
or configurations  depending on the plot
function of this package. The output is a data frame.
}
\description{
function compares matrices of solutions
to prepare the data for charting with the UpSetR
}
