% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions_robust.R
\name{conditions_robust}
\alias{conditions_robust}
\title{Robustness for single (INUS) conditions}
\usage{
conditions_robust(ls, nsets = 5)
}
\arguments{
\item{ls}{List of QCA models as produced
with \code{\link[QCA-minimize]{QCA::minimize()}} from \pkg{QCA}
package}

\item{nsets}{Specifies number of sets to be plotted. Argument
imported from the \code{\link{upset}} function from \pkg{UpSetR}.}
}
\value{
A plot presenting the frequency of individual
conditions and their co-occurrences across QCA models.
}
\description{
Evaluates robustness on the level of single conditions
or INUS conditions (Insufficient and Necessary conditions
that are part of a conjunction that is Unnecessary and
Sufficient). Aggregates over conditions by calculating
and plotting their frequencies and the frequencies of
conditions that co-occur in a QCA model
}
\examples{
data("solutions_example")
conditions_robust(solutions_example)

}
