% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conds_robust.R
\name{conds_robust}
\alias{conds_robust}
\title{conds_robust()}
\usage{
conds_robust(df, nsets)
}
\arguments{
\item{df}{Dataframe or list of QCA solutions or configurations
derived from multiple truth table analyses performed
with \code{\link[QCA]{minimize}} from \code{\link{QCA}}
package}

\item{nsets}{Specifies number of sets to be plotted. Argument
imported from the \code{\link{upset}} function from \pkg{UpSetR}.}
}
\value{
A plot presenting the frequency of individual
conditions and their co-occurrences across QCA solutions.
}
\description{
Evaluates robustness on the level of single conditions
or INUS conditions (Insufficient and Necessary conditions
that are part of a conjunction that is Unnecessary and
Sufficient). It aggregates over conditions by calculating
and plotting their frequencies and the frequencies of
conditions cooccurring in a solution.
}
\examples{
sols <- list(c(replicate("A", 2), replicate("AB+c", 2), replicate("A+c", 2)))
conds_robust(sols)

}
