% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCA_viz.R
\name{conds_upset}
\alias{conds_upset}
\title{conds_upset()}
\usage{
conds_upset(df, nsets)
}
\arguments{
\item{df}{the data frame which has the
configurations extracted from the QCA solutions only}

\item{nsets}{an argument imported from the UpSetR
package. Determines the number of sets to graph.}
}
\value{
The function returns an UpSetR-generated
visual represenation of the interconnecteds
between the conditions in a given dataset
}
\description{
Plot Function to chart conditions
intersection of the QCA output in a solution.
}
\note{
the chart generated by this function is
different from that generated by the sols_upset
}
