% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCA_viz.R
\name{conds_upset_h}
\alias{conds_upset_h}
\title{conds_upset_h()}
\usage{
conds_upset_h(df, nsets)
}
\arguments{
\item{df}{is an object that contains hand-pooled (concatenated)
solutions produced by the \pkg{QCA} package.
This object includes only the solutions.}

\item{nsets}{an argument imported from the \pkg{UpSetR}
package. Determines the number of sets to graph.}

\item{type}{an argument to specify the panel dimension.
One can specify either within, between or
pooled types. Input is a string and there is
no default option for this argument.}
}
\value{
The function returns an UpSetR-generated
visual represenation of the interconnecteds
between the conditions in a given dataset
}
\description{
Plot Function for the charting of conditions
intersection of the panel QCA output in a solution.
}
\note{
the chart generated by this function is
different from that generated by the sols_upset
}
