% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCA_viz.R
\name{config_panel}
\alias{config_panel}
\title{config_panel()}
\usage{
config_panel(df, nsets, type)
}
\arguments{
\item{df}{the data frame which has the
configurations extracted from the QCA solutions only}

\item{nsets}{an argument imported from the UpSetR
package. Determines the number of sets to graph.}

\item{type}{an argument to specify the panel dimension.
One can specify either within, between or
pooled types. Input is a string and there is 
no default option for this argument.}
}
\value{
The function returns an UpSetR-generated
visual represenation of the interconnecteds
between the conditions in a given dataset
}
\description{
Plot Function for the charting of configuration
intersection of the panel QCA output in a solution.
}
\note{
the chart generated by this function is
different from that generated by the sols_upset
}
