% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCA_viz.R
\name{config_upset}
\alias{config_upset}
\title{config_upset()}
\usage{
config_upset(df, nsets)
}
\arguments{
\item{df}{Dataframe with configurations extracted
from the QCA solutions}

\item{nsets}{An argument imported from the \pkg{UpSetR}
package. Determines the number of sets to be plotted.}
}
\value{
An \pkg{UpSetR}-generated intersection plot
presenting the frequency of individual configurations across
QCA solutions and their intersections across solutions.
}
\description{
Function decomposes results into individual configurations
and creates intersection plots.
}
