% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCA_viz.R
\name{config_upset_h}
\alias{config_upset_h}
\title{config_upset_h()}
\usage{
config_upset_h(df, nsets)
}
\arguments{
\item{df}{the data frame which has the
configurations extracted from the QCA solutions only}

\item{nsets}{an argument imported from the UpSetR
package. Determines the number of sets to graph.}

\item{type}{an argument to specify the panel dimension.
One can specify either within, between or
pooled types. Input is a string and there is
no default option for this argument.}
}
\value{
The function returns an UpSetR-generated
visual represenation of the interconnecteds
between the conditions in a given dataset
}
\description{
config_upset_h()
}
\note{
the present version of config_upset
does not allow for selection of thresholds
as opposed to it's sibling function config_upset
}
