% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_robust.R
\name{configs_robust}
\alias{configs_robust}
\title{configs_robust()}
\usage{
configs_robust(df, const = FALSE, nsets)
}
\arguments{
\item{df}{Dataframe of QCA solutions as produced
with \code{\link[QCA]{minimize}} from \code{\link{QCA}}
package}

\item{const}{Choose between the plotting of all solutions
(\code{const = F} or only those meeting a minimum consistency
threshold specified with \code{const}.}

\item{nsets}{Specifies number of sets to be plotted. Argument
imported from the \code{\link{upset}} function from \pkg{UpSetR}.}

\item{y}{is an argument with default value of zero. If
any other number is set for the given argument, only
those configurations which are above the set value for y
will be plotted by the function}
}
\value{
A plot presenting the frequency of individual
conditions and their co-occurrences across QCA solutions.
}
\description{
Evaluates robustness on the level of individual
sufficient conditions or conjunctions (terms). It aggregates
over sufficient terms by calculating their frequencies and
the frequencies of terms cooccurring in a solution. The
frequencies are plotted using Upset plots.
}
