% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations_robust.R
\name{configurations_robust}
\alias{configurations_robust}
\title{Robustness for sufficient terms (usually configurations)}
\usage{
configurations_robust(ls, const = FALSE, nsets = 5)
}
\arguments{
\item{ls}{List of QCA models as produced
with \code{\link[QCA]{minimize}} from \code{\link{QCA}}
package}

\item{const}{Choose between the plotting of all solutions
(\code{const = FALSE} or only those meeting a minimum consistency
threshold specified with \code{const}.}

\item{nsets}{Specifies number of sets to be plotted. Argument
imported from the \code{\link{upset}} function from \pkg{UpSetR}.}

\item{y}{is an argument with default value of zero. If
any other number is set for the given argument, only
those configurations which are above the set value for y
will be plotted by the function.}
}
\value{
A plot presenting the frequency of individual
terms and their co-occurrences across QCA models
}
\description{
Evaluate robustness on the level of individual
sufficient terms (conditions or conjunctions/configurations) of models.
Aggregates over sufficient terms by calculating
their frequencies and the frequencies of terms co-occurring
in a model. The frequencies are plotted using Upset plots from
\code{\link{UpSetR}}.
}
\examples{
data("solutions_example")
# analysis without any consistency threshold
configurations_robust(solutions_example)

# analysis with consistency >= .75
configurations_robust(KCP_example, const = 0.75)

}
