% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCArobust.R, R/cons_minimum.R
\name{cons_minimum}
\alias{cons_minimum}
\title{Preselecting configurations for analysis based on minimum consistency value}
\usage{
cons_minimum(x, conthresh = 0)

cons_minimum(x, conthresh = 0)
}
\arguments{
\item{x}{output of the QCA in its entirety
produced by \pkg{QCA} package}

\item{con.thresh}{Minimum consistency threshold for being included in
the analysis. Default is zero.}
}
\value{
A list of a subset of configurations meeting the threshold.

A list of a subset of configurations meeting the threshold.
}
\description{
Preselecting configurations for analysis based on minimum consistency value

Preselecting configurations for analysis bsaed on minimum consistency value
}
