% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_minimum.R
\name{cov_minimum}
\alias{cov_minimum}
\title{Filtering configurations based on coverage threshold.}
\usage{
cov_minimum(x, covthresh)
}
\arguments{
\item{x}{List of QCA solutions or configurations
derived from multiple truth table analyses performed
with \code{\link[QCA]{minimize}} from \code{\link{QCA}}
package}

\item{covthresh}{Coverage threshold for filtering
solutions or configurations. Default value is 0.}
}
\value{
A dataframe with configurations meeting the coverage
threshold.
}
\description{
Filtering configurations based on coverage threshold.
}
