% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCA_viz.R
\name{detection}
\alias{detection}
\title{detection()}
\usage{
detection(df, x)
}
\arguments{
\item{df}{list of all the solutions or configurations
which are obtained from multiple QCA solutions}

\item{x}{QCA solutions in their raw form as produced
by \pkg{QCA} package}
}
\value{
The function counts the individual solutions
or configurations  depending on the plot
function of this package. The output is a data frame.
}
\description{
a new version of comparison function designed for the
configurations. Needs further reworking to get rid
of the for loop at later stages
}
