% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCArobust.R
\name{dt.selector}
\alias{dt.selector}
\title{dt.selector()}
\usage{
dt.selector(x, con.thresh = 0)
}
\arguments{
\item{x}{output of the QCA in its entirety
produced by \pkg{QCA} package}

\item{con.thresh}{is the threshold set by a
researcher. Te default value equals to zero}
}
\value{
The function returns a subset of
configurations that are larger than the selected
threshold value.
}
\description{
function that allows for selecting configurations
bsaed on consistency values before plitting
}
