% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.selector2.R
\name{dt.selector2}
\alias{dt.selector2}
\title{Filtering configurations based on coverage threshold.}
\usage{
dt.selector2(x, covthresh)
}
\arguments{
\item{x}{List of QCA solutions or configurations
derived from multiple truth table analyses performed
with \code{\link[QCA]{minimize}} from \code{\link{QCA}}
package}

\item{convthresh}{Coverage threshold for filtering
solutions or configurations. Default value is 0.}
}
\value{
A dataframe with configurations meeting the coverage
threshold.
}
\description{
Filtering configurations based on coverage threshold.
}
