% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCA_viz.R
\docType{data}
\name{hinter}
\alias{hinter}
\title{Hinter-Leitner dataset
Data borrowed from Hinter and Leitner study
A dataset containing informaiton about 20 countres
used to understand the politic sof external
approval: Explaining the IMF's evaluation
of austerity programmes in those countries.}
\format{A list with 20 observations of 10 variables
\describe{
  \item{case}{the case covered in the study. The country being analyzed}
  \item{pos}{the numeric value of pos}
  \item{reduction}{the numeric value of reduction}
  \item{duration}{the numeric value of duration}
  \item{ambition}{the numeric value of ambition}
  \item{DC}{the numeric value of DC}
  \item{EFF}{the numeric value of EFF}
  \item{CR}{the numeric value of CR}
  \item{COM}{the numeric value of COM}
  \item{PRG}{the numeric value of PRG}
}}
\source{
\url{http://onlinelibrary.wiley.com/doi/10.1111/1475-6765.12142/suppinfo/}

\url{https://www.cambridge.org/core/journals/political-analysis/article/fuzzy-sets-on-shaky-ground-parameter-sensitivity-and-confirmation-bias-in-fsqca/B693F136C5158DF3C19686EC89522C23}
}
\usage{
hinter

hinter
}
\description{
Hinter-Leitner dataset
Data borrowed from Hinter and Leitner study
A dataset containing informaiton about 20 countres
used to understand the politic sof external
approval: Explaining the IMF's evaluation
of austerity programmes in those countries.

Krogslund dataset
Data borrowed from Krogslund/Choi/Poertner study
A dataset containing informaiton about 5 variables
that are the result of a simulation
}
\keyword{datasets}
