% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCA_viz.R
\name{reduced_c}
\alias{reduced_c}
\title{reduced_c()
a wrapper function to minimize causal conditions of the
QCA truth table for a parsimonious solution}
\usage{
reduced_c(df)
}
\arguments{
\item{df}{truth table derived with the truth function}

\item{details}{an argument, asking wheter to pring
the details of a solution or not}

\item{all.sol}{an argument for printing of all solutions
irrespective of the number of prime implicants}

\item{include}{other output data to be included while
minimizing the table}

\item{dir.exp}{an argument to defin directional
expectations for conserveative solutions}
}
\description{
reduced_c()
a wrapper function to minimize causal conditions of the
QCA truth table for a parsimonious solution
}
