% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCA_viz.R
\name{reduced_int}
\alias{reduced_int}
\title{reduced_int()}
\usage{
reduced_int(df, dir)
}
\arguments{
\item{df}{truth table derived with the truth function}

\item{details}{an argument, asking whether to print
the details of a solution or not}

\item{all.sol}{an argument for printing of all solutions
irrespective of the number of prime implicants}

\item{include}{other input data to be included when
minimizing the table}

\item{dir.exp}{an argument to define directional expectations
for intermediate solutions}
}
\value{
returns a particular data type
}
\description{
a wrapper function to minimize causal conditions
of the QCA truth table for an intermediate solution
}
\note{
this is a wrapper function
}
