% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCA_viz.R
\name{reduced_p}
\alias{reduced_p}
\title{reduced_p()}
\usage{
reduced_p(df)
}
\arguments{
\item{df}{truth table derived with the truth_function}

\item{details}{an argument, asking whether to
print the details of a solution or not}

\item{all.sol}{an argument for printing of all solutions
irrespective of the number of prime implicants}

\item{include}{other output data to be included while
minimizing the table}

\item{dir.exp}{an argument to define directional
expectations for parsimonious solutions}
}
\value{
returns a QCA-specific data type
}
\description{
a wrapper function to minimize causal conditions
of the QCA truth table for a parsimonious solution
}
\note{
this is a wrapper function
}
