% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Skaaning_table3}
\alias{Skaaning_table3}
\title{Skaaning dataset
Data is borrowed from the robustness analysis by Skaaning and Sven-Erik (2011)
The list of solutions presented in table 3 of the article. It  summarizes
results for the choice of different calibration anchors for crisp-set QCA.}
\format{
A list of 244 solutions.
\describe{
  \item{solution1}{First solution}
  \item{solution2}{Second solution}
  \item{solution3}{Third solution}
  \item{solution4}{Fourth solution}
  \item{solution5}{Fifth solution}
}
}
\source{
\url{https://doi.org/10.1177/0049124111404818}
}
\usage{
Skaaning_table3
}
\description{
Skaaning dataset
Data is borrowed from the robustness analysis by Skaaning and Sven-Erik (2011)
The list of solutions presented in table 3 of the article. It  summarizes
results for the choice of different calibration anchors for crisp-set QCA.
}
\keyword{datasets}
