% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sols_aggregate.R
\name{sols_aggregate}
\alias{sols_aggregate}
\title{sols_aggregate()}
\usage{
sols_aggregate(ls, plot = T, plot_solutions = 5)
}
\arguments{
\item{ls}{List of QCA solutions or configurations
derived from multiple truth table analyses performed
with \code{\link[QCA]{minimize}} from \pkg{QCA}
package.}

\item{plot}{If set to \code{TRUE}, a bar chart of frequency
of solutions is produced. If set to \code{FALSE},
a dataframe with solutions and frequencies is returned.
Default is \code{TRUE}.}

\item{plot_solutions}{Number of solutions to be plotted in
bar chart. Default is 5. If \code{plot} is set to \code{FALSE},
the dataframe shows all solutions regardless of the number
specified for plotting.}
}
\value{
If \code{plot} set to \code{TRUE}, a bar chart of
frequency of solutions is produced with \code{\link{ggplot2}}.
If \code{plot} is set to \code{FALSE}, a dataframe with
solutions and frequencies.
}
\description{
Aggregates over solutions by calculating and plotting
their frequencies.
}
