% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCA_viz.R
\name{sols_barplot}
\alias{sols_barplot}
\title{sols_barplot()}
\usage{
sols_barplot(ls)
}
\arguments{
\item{ls}{is a list of all solutions in their
entireity and row format}
}
\value{
the function returns a frequency distirbution
of the solutions as they are in their entirety
}
\description{
Frequency plotting of the solutoins
}
