% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCA_viz.R
\name{sols_panel}
\alias{sols_panel}
\title{sols_panel()}
\usage{
sols_panel(df, nsets, type)
}
\arguments{
\item{df}{the data frame which has the
configurations extracted from the QCA output
in its entirety}

\item{nsets}{an argument imported from the UpSetR
package. Determines the number of sets to graph.}

\item{type}{an argument to specify the panel dimension.
One can specify either within, between or
pooled types. Input is a string and there is 
no default option for this argument.}
}
\value{
The function returns an UpSetR-generated
visual represenation of the interconnecteds
between the conditions in a given dataset
}
\description{
Plot Function for the charting of solutions
intersections of the panel QCA output.
}
\note{
this chart is different from that which
is generated by conds_upset
}
