% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sols_plot.R
\name{sols_plot}
\alias{sols_plot}
\title{sols_plot()}
\usage{
sols_plot(ls)
}
\arguments{
\item{ls}{List of QCA solutions or configurations
derived from multiple truth table analyses performed
with \code{\link[QCA]{minimize}} from \pkg{QCA}
package.}
}
\value{
A bar chart created with \code{\link{ggplot2}} summarizing
the frequency of solutions in the data.
}
\description{
Plots the frequency of solutions using a bar chart.
}
