% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCA_viz.R
\name{sols_upset}
\alias{sols_upset}
\title{sols_upset()}
\usage{
sols_upset(df, nsets)
}
\arguments{
\item{df}{the data frame which has the
configurations extracted from the QCA output
in its entirety}

\item{nsets}{an argument imported from the UpSetR
package. Determines the number of sets to graph.}
}
\value{
The function returns an UpSetR-generated
visual represenation of the interconnecteds
between the conditions in a given dataset
}
\description{
Plot Function for the charting of configuration
intersection of the QCA output
}
\note{
this chart is different from that which
is generated by conds_upset
}
