% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions_robust.R
\name{solutions_robust}
\alias{solutions_robust}
\title{Evaluate robustness of complete QCA models}
\usage{
solutions_robust(ls, plot = TRUE, plot_solutions = 5)
}
\arguments{
\item{ls}{List of QCA solutions or configurations
derived from multiple truth table analyses performed
with \code{\link[QCA]{minimize}} from \pkg{QCA}
package.}

\item{plot}{If set to \code{TRUE}, a bar chart of frequency
of solutions is produced. If set to \code{FALSE},
a dataframe with solutions and frequencies is returned.
Default is \code{TRUE}.}

\item{plot_solutions}{Number of solutions to be plotted in
bar chart. Default is 5. If \code{plot} is set to \code{FALSE},
the dataframe shows all solutions regardless of what number has been
specified.}
}
\value{
If \code{plot} set to \code{TRUE}, a bar chart of
frequency of solutions is produced with \code{\link{ggplot2}}.
If \code{plot} is set to \code{FALSE}, a dataframe with
solutions and frequencies.
}
\description{
Aggregate over multiple solutions and calculate and plot
model frequencies.
}
\examples{
data("Skaaning_table3")
# plotting models with default setting
solutions_robust(Skaaning_table3, plot = TRUE)
# plotting three solutions
solutions_robust(Skaaning_table3, plot = TRUE, plot_solutions = 3)

}
