% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCA_viz.R
\name{solutions_table}
\alias{solutions_table}
\title{solutions_table()}
\usage{
solutions_table(ls)
}
\arguments{
\item{ls}{is a list of all solutions in their
entireity and row format}
}
\value{
the function returns a table of solutions
and their frequency of occurrence in a desceinding
order (in the ready for rmarkwdon)
}
\description{
Plot Function for the table version of solutions
the special  function which allows to
chart the solutions in their entirety inside
a table which is transferrable to R markdown
}
