% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCA_viz.R
\name{solutions_table_S}
\alias{solutions_table_S}
\title{solutions_table_S()}
\usage{
solutions_table_S(ls)
}
\arguments{
\item{ls}{is a list of all solutions in their
entireity and row format}
}
\value{
the function returns a table of solutions
and their frequency of occurrence in a desceinding
order (in the ready for rmarkwdon)
}
\description{
Plot Function for the table version of solutions
the present function is built for data that is
inputted manually in a form of a non-QCA list.
}
