% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCA_viz.R
\name{comparison}
\alias{comparison}
\title{comparison()}
\usage{
comparison(x = all_values, y, num = F)
}
\arguments{
\item{x}{list of all the solutions or configurations
which are obtained form the QCA solutions}

\item{y}{QCA solutions in their raw form}

\item{num}{is a default argument which checks
whether the inputs are numeric or not. The default
is set for false. This ensures that even if the input
is non-numeric, the function will read in the
data in a numeric format.}
}
\value{
The function counts the configurations
and/or individual solutions depending on the plot
funciton used later. The output is presented in
the form of a data frame
}
\description{
function compares matrices of solutions
to prepare the data for charting with the UpSetR
}
