% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{comparison}
\alias{comparison}
\title{Comparison of matrices preparing data for plotting}
\usage{
comparison(x = all_values, y, num = F)
}
\arguments{
\item{x}{List of QCA solutions or configurations
derived from multiple truth table analyses performed
with \code{\link{QCA}} package}

\item{y}{QCA solutions as produced
with \code{\link[QCA]{minimize}} from \code{\link{QCA}}
package}

\item{num}{Check for numeric input. Default
is set to \code{FALSE}. This setting ensures that the function
will read data in a numeric format even it is in a
different format.}
}
\value{
A dataframe counting the individual solutions
or configurations.
}
\description{
Comparison of matrices preparing data for plotting
}
