% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCArobust.R
\name{comparison}
\alias{comparison}
\title{Comparison of matrices preparing data for plotting (internal function)}
\usage{
comparison(x = all_values, y, num = F)
}
\arguments{
\item{x}{List of QCA solutions or configurations
derived from multiple truth table analyses}

\item{y}{QCA solutions in their raw form as produced
by \code{\link[QCA]{minimize}} from \code{QCA} package}

\item{num}{Checks whether the input is  numeric or not. The default
is set to \code{FALSE}. This ensures that even if the input
is non-numeric, the function will read in the
data in a numeric format.}
}
\value{
A dataframe counting the individual solutions
or configurations.
}
\description{
Comparison of matrices preparing data for plotting (internal function)
}
