% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions_robust.R
\name{conditions_robust}
\alias{conditions_robust}
\title{Robustness for single (INUS) conditions}
\usage{
conditions_robust(df, nsets = 5)
}
\arguments{
\item{df}{Dataframe or list of QCA solutions or configurations
derived from multiple truth table analyses performed
with \code{\link[QCA]{minimize}} from \code{\link{QCA}}
package}

\item{nsets}{Specifies number of sets to be plotted. Argument
imported from the \code{\link{upset}} function from \pkg{UpSetR}.}
}
\value{
A plot presenting the frequency of individual
conditions and their co-occurrences across QCA models.
}
\description{
Evaluates robustness on the level of single conditions
or INUS conditions (Insufficient and Necessary conditions
that are part of a conjunction that is Unnecessary and
Sufficient). It aggregates over conditions by calculating
and plotting their frequencies and the frequencies of
conditions that coincide in a QCA model
}
\examples{
sols <- list(c(rep("A", 2), rep("AB+c", 2), rep("A+c", 3)))
conditions_robust(sols)

}
