% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCA_viz.R
\name{conds_upset}
\alias{conds_upset}
\title{conds_upset()}
\usage{
conds_upset(df, nsets)
}
\arguments{
\item{df}{Dataframe with solutions extracted
from the solutions of the \pkg{QCA} package}

\item{nsets}{An argument imported from the \pkg{UpSetR}
package. Determines the number of sets to be plotted.}
}
\value{
An \pkg{UpSetR}-generated intersection plot
presenting the frequency of individual conditions across
QCA solutions and their intersections across solutions.
}
\description{
Function decomposes results into individual conditions
and creates intersection plots.
}
