% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conds_upset.R
\name{conds_upset}
\alias{conds_upset}
\title{Function decomposes results into individual conditions
and creates intersection plots.}
\usage{
conds_upset(df, nsets)
}
\arguments{
\item{df}{Dataframe or list of QCA solutions or configurations
derived from multiple truth table analyses performed
with \code{\link[QCA]{minimize}} from \code{\link{QCA}}
package}

\item{nsets}{Specifies number of sets to be plotted. Argument
imported from the \code{\link{upset}} function from \pkg{UpSetR}.}
}
\value{
A plot presenting the frequency of individual
conditions and their co-occurrences across QCA solutions.
}
\description{
Function decomposes results into individual conditions
and creates intersection plots.
}
\examples{
sols <- list(c(replicate("A", 2), replicate("AB+c", 2), replicate("A+c", 2)))
conds_upset(sols)

}
