% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conds_upset.R
\name{conds_upset}
\alias{conds_upset}
\title{Function decomposes results into individual conditions
and creates intersection plots.}
\usage{
conds_upset(df, nsets)
}
\arguments{
\item{df}{Dataframe or list with solutions extracted
from the solutions of the \code{\link{QCA}} package}

\item{nsets}{Argument imported from the \code{\link{upset}}
function from \pkg{UpSetR}. Specifies number of sets to be plotted.}
}
\value{
A plot presenting the frequency of individual
conditions and their co-occurrences across QCA solutions.
}
\description{
Function decomposes results into individual conditions
and creates intersection plots.
}
\examples{
sols <- list(c(replicate("A", 2), replicate("AB+c", 2), replicate("A+c", 2)))
conds_upset(sols)

}
