% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCA_viz.R
\name{config_upset}
\alias{config_upset}
\title{config_upset()}
\usage{
config_upset(df, nsets)
}
\arguments{
\item{df}{the data frame which has the
configurations extracted from the QCA solutions only.}

\item{nsets}{an argument imported from the UpSetR
package. Determines the number of sets to graph.}
}
\value{
The function returns an UpSetR-generated
visual represenation of the interconnecteds
between the conditions in a given dataset
}
\description{
Plot Function for the charting of configurations'
intersection of the QCA output in a solution.
}
\note{
the chart generated by this function is
different from that generated by the sols_upset
}
