% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cons_minimum.R
\name{cons_minimum}
\alias{cons_minimum}
\title{Filtering configurations from results using minimum consistency value}
\usage{
cons_minimum(x, consthresh = 0)
}
\arguments{
\item{x}{List of solutions generated with \code{\link{QCA}} package}

\item{consthresh}{Minimum consistency threshold for for filtering
configurations. Default is zero.}
}
\value{
A dataframe with configurations meeting the consistency
threshold.
}
\description{
Filtering configurations from results using minimum consistency value
}
