% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\name{counting}
\alias{counting}
\title{Counts frequency of a term in list of results}
\usage{
counting(list, term)
}
\arguments{
\item{list}{List of terms that is searched for a match}

\item{term}{Term (single condition or conjunction) that should
be searched for}
}
\value{
Count of search terms in the list that has been searched
}
\description{
Counts frequency of a term in list of results
}
\examples{
conjunction_list <- list(c("ABC", "ACD", "ABD"))
counting(conjunction_list, "AC")
counting(conjunction_list, "AB")

}
