% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCA_viz.R
\name{counting}
\alias{counting}
\title{counting()}
\usage{
counting(list, terms)
}
\arguments{
\item{list}{the data from within which we want to count}

\item{terms}{are the terms that we need to count}
}
\value{
returns a the count of terms within the list of strings
}
\description{
Function that counts a specific string combination
}
\note{
this functiion uses the stri_count function of the
stringi function to count  a pattern in a list of strings
}
