% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCArobust.R, R/counting.R
\name{counting}
\alias{counting}
\title{counting()}
\usage{
counting(list, term)

counting(list, term)
}
\arguments{
\item{list}{List of terms that is searched for a match}

\item{term}{Term (single condition or conjunction) that should
be searched for}

\item{terms}{are the terms that we need to count}
}
\value{
returns a the count of terms within the list of strings

Count of search terms in the list that has been searched
}
\description{
Function that counts a specific string combination
}
\note{
this functiion uses the stri_count function of the
stringi function to count  a pattern in a list of strings
}
\examples{
conjunction_list <- list(c("ABC", "ACD", "ABD"))
counting(conjunction_list, "AC")
counting(conjunction_list, "AB")

}
