% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detection}
\alias{detection}
\title{Version of comparison function for configurations.}
\usage{
detection(df, x)
}
\arguments{
\item{df}{List of QCA solutions or configurations
derived from multiple truth table analyses performed
with \code{\link[QCA]{minimize}} from \code{\link{QCA}}
package}
}
\value{
A dataframe counting the individual solutions
or configurations.
}
\description{
Version of comparison function for configurations.
}
