% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detection}
\alias{detection}
\title{Version of comparison function for configurations.}
\usage{
detection(ls, x)
}
\arguments{
\item{ls}{List of QCA solutions or configurations
derived from multiple truth table analyses performed
with \code{\link[QCA]{minimize}} from \pkg{QCA}
package}

\item{x}{a vector of all unique values derived from the
list of QCA solutions}
}
\value{
A list counting the individual solutions
or configurations.
}
\description{
Version of comparison function for configurations.
}
