% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCArobust.R, R/detection.R
\name{detection}
\alias{detection}
\title{New version of comparison function designed for the
configurations. Needs further reworking to get rid
of the for loop at later stages}
\usage{
detection(df, x)

detection(df, x)
}
\arguments{
\item{df}{list of all the solutions or configurations
which are obtained from multiple QCA solutions}

\item{x}{QCA solutions in their raw form as produced
by \pkg{QCA} package}
}
\value{
The function counts the individual solutions
or configurations  depending on the plot
function of this package. The output is a data frame.

The function counts the individual solutions
or configurations  depending on the plot
function of this package. The output is a data frame.
}
\description{
New version of comparison function designed for the
configurations. Needs further reworking to get rid
of the for loop at later stages

New version of comparison function designed for the
configurations. Needs further reworking to get rid
of the for loop at later stages
}
