% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.selector2.R
\name{dt.selector2}
\alias{dt.selector2}
\title{dt.selector2()}
\usage{
dt.selector2(x, cov.thresh)
}
\arguments{
\item{x}{output of the QCA in its entirety
produced by \pkg{QCA} package}

\item{con.thresh}{is the threshold set by a
researcher. Te default value equals to zero}
}
\value{
The function returns a subset of
configurations that are larger than the selected
threshold value.
}
\description{
function that allows for selecting specific
coverage threshold.
}
