% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KCP_model1_ncut1}
\alias{KCP_model1_ncut1}
\title{Datasets documentation
Krogslund dataset (model 1, ncut 1)
Data borrowed from Krogslund/Choi/Poertner study
A dataset containing informaiton about 5 variables
that are the result of a simulation analysis. This
dataset refers to a simulation analysis of original data
by Koenig-Archibugi.}
\format{
A tibble with 40000 observations and 5 variables
\describe{
  \item{incl.cut1.val}{First inclusion cut}
  \item{incl.cut0.val}{Second inclusion cut}
  \item{n.cut.val}{the numeric value}
  \item{error}{the error}
  \item{Configurations}{Configurations vector}
}
}
\source{
\url{https://doi.org/10.1093/pan/mpu016}
}
\usage{
KCP_model1_ncut1
}
\description{
Datasets documentation
Krogslund dataset (model 1, ncut 1)
Data borrowed from Krogslund/Choi/Poertner study
A dataset containing informaiton about 5 variables
that are the result of a simulation analysis. This
dataset refers to a simulation analysis of original data
by Koenig-Archibugi.
}
\keyword{datasets}
