% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCA_viz.R
\name{sols_upset}
\alias{sols_upset}
\title{sols_upset()}
\usage{
sols_upset(df, nsets)
}
\arguments{
\item{df}{Dataframe containing multiple QCA solutions.}

\item{nsets}{An argument imported from the \pkg{UpSetR}
package. Determines the number of sets to be plotted.}
}
\value{
A bar chart with frequencies for multiple solutions.
}
\description{
Function plots frequencies of multiple QCA solutions.
}
