% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{solutions_example}
\alias{solutions_example}
\title{Datasets documentation
List of hypothetical solutions}
\format{
A list with 10 solutions produced with \code{minimize} from the
\pkg{QCA} package.
}
\usage{
solutions_example
}
\description{
A list with ten hypothetical solutions derived from a truth table analysis.
}
\keyword{datasets}
