% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions_robust.R
\name{solutions_robust}
\alias{solutions_robust}
\title{Robustness for complete QCA models}
\usage{
solutions_robust(ls, plot = TRUE, plot_solutions = 5)
}
\arguments{
\item{ls}{List of QCA solutions
derived from multiple truth table analyses performed
with \code{\link[QCA]{minimize}} from \pkg{QCA}
package.}

\item{plot}{If set to \code{TRUE}, one produces a bar chart of frequency
of models. If set to \code{FALSE},
a dataframe with models and their frequencies is returned.
Default is \code{TRUE}.}

\item{plot_solutions}{Number of solutions to be plotted in
bar chart. Default is 5. If \code{plot} is set to \code{FALSE},
the dataframe shows all solutions regardless of what number one has
specified.}
}
\value{
If \code{plot} set to \code{TRUE}, a bar chart of
frequency of solutions is produced with \code{\link{ggplot2}} that can be
customized. If \code{plot} is set to \code{FALSE}, a dataframe with
solutions and frequencies is returned that can be further processed.
}
\description{
Aggregate over multiple model and calculate and plot
model frequencies.
}
\examples{
data("solutions_example")
# plotting models with default setting
solutions_robust(solutions_example, plot = TRUE)
# plotting three solutions
solutions_robust(solutions_example, plot = TRUE, plot_solutions = 3)

}
